Refactoring & client-side rendering (#1130)
