+++ monitoring/ingest_inspector/frontend/next.config.mjs
@@ -1,4 +1,6 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+    output: 'export',
+};
 
 export default nextConfig;
